requirements.txt
bleak==1.0.1
google-cloud-firestore
google-auth
